# form/openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "description": "A local user account present on the host.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "kind": {
            "const": "account",
            "default": "account",
            "title": "Kind",
            "type": "string"
          },
          "last_login": {
            "anyOf": [
              {
                "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "shell": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shell"
          },
          "uid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uid"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "username"
        ],
        "title": "Account",
        "type": "object"
      },
...
      "AssetReport": {
        "description": "agentd -> Form -> analyzer: one host, one collection cycle.",
        "properties": {
          "assets": {
            "items": {
              "discriminator": {
                "mapping": {
                  "account": "#/components/schemas/Account",
                  "container": "#/components/schemas/Container",
                  "credential": "#/components/schemas/Credential",
                  "image": "#/components/schemas/Image",
                  "package": "#/components/schemas/Package",
                  "port": "#/components/schemas/Port",
                  "security_product": "#/components/schemas/SecurityProduct",
                  "service": "#/components/schemas/Service"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Package"
                },
                {
                  "$ref": "#/components/schemas/Service"
                },
                {
                  "$ref": "#/components/schemas/Port"
                },
                {
                  "$ref": "#/components/schemas/Account"
                },
                {
                  "$ref": "#/components/schemas/Credential"
                },
                {
                  "$ref": "#/components/schemas/Container"
                },
                {
                  "$ref": "#/components/schemas/Image"
                },
                {
                  "$ref": "#/components/schemas/SecurityProduct"
                }
              ]
            },
            "maxItems": 4096,
            "title": "Assets",
            "type": "array"
          },
          "collected_at": {
            "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "detector_runs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetectorRun"
                },
                "maxItems": 32,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detectors explicitly executed by the producer. null means legacy/unknown; an empty list means the producer confirms none were enabled.",
            "title": "Detector Runs"
          },
          "host": {
            "$ref": "#/components/schemas/HostInfo"
          },
          "report_id": {
            "maxLength": 256,
            "title": "Report Id",
            "type": "string"
          },
          "scanner_version": {
            "maxLength": 256,
            "title": "Scanner Version",
            "type": "string"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticated Agent identity injected by Form; never trusted from payload",
            "title": "Source Agent Id"
          },
          "source_target_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Form-owned registered target attribution; absent only for unbound legacy telemetry",
            "title": "Source Target Id"
          },
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/Vulnerability"
            },
            "maxItems": 4096,
            "title": "Vulnerabilities",
            "type": "array"
          }
        },
        "required": [
          "report_id",
          "collected_at",
          "scanner_version",
          "host"
        ],
        "title": "AssetReport",
        "type": "object"
      },
...
      "Container": {
        "description": "A container workload discovered from static runtime metadata.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to the static metadata file under scan_root",
            "title": "Config Path"
          },
          "container_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Runtime container id when available",
            "title": "Container Id"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Image reference when known from static metadata",
            "title": "Image"
          },
          "kind": {
            "const": "container",
            "default": "container",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "rootfs_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Merged container rootfs path under scan_root when resolved statically",
            "title": "Rootfs Path"
          },
          "runtime": {
            "description": "Container runtime, e.g. docker / podman / containerd / kubernetes",
            "title": "Runtime",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last known state, e.g. running / exited / created",
            "title": "Status"
          }
        },
        "required": [
          "asset_id",
          "name",
          "runtime"
        ],
        "title": "Container",
        "type": "object"
      },
      "Credential": {
        "description": "A credential artifact found on the host, referenced only by its public fingerprint.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "credential_kind": {
            "$ref": "#/components/schemas/CredentialKind"
          },
          "fingerprint": {
            "description": "Public fingerprint or hash; the secret itself MUST NEVER be transmitted",
            "title": "Fingerprint",
            "type": "string"
          },
          "kind": {
            "const": "credential",
            "default": "credential",
            "title": "Kind",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "required": [
          "asset_id",
          "credential_kind",
          "fingerprint"
        ],
        "title": "Credential",
        "type": "object"
      },
...
      "CredentialKind": {
        "description": "Kind of credential material discovered on a host.",
        "enum": [
          "ssh_key",
          "api_key",
          "password",
          "token"
        ],
        "title": "CredentialKind",
        "type": "string"
      },
...
      "DetectorRun": {
        "description": "Agent-declared execution evidence; absence differs from a zero finding run.",
        "properties": {
          "detector": {
            "$ref": "#/components/schemas/DetectorKind"
          },
          "finding_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Finding Count",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/DetectorRunStatus",
            "default": "complete"
          }
        },
        "required": [
          "detector"
        ],
        "title": "DetectorRun",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Image": {
        "description": "A container image present in local runtime storage (a pulled image that\nmay never have run), discovered from static on-disk metadata.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "created": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Image creation time from the image config, when available",
            "title": "Created"
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Content-addressable image id (e.g. sha256:...)",
            "title": "Image Id"
          },
          "kind": {
            "const": "image",
            "default": "image",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "description": "Primary image reference (e.g. nginx:1.25), or short image id when untagged",
            "title": "Name",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "runtime": {
            "description": "Image store / runtime, e.g. docker / podman",
            "title": "Runtime",
            "type": "string"
          },
          "tags": {
            "description": "All repository tags / names for this image",
            "items": {
              "type": "string"
            },
            "maxItems": 256,
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "asset_id",
          "name",
          "runtime"
        ],
        "title": "Image",
        "type": "object"
      },
...
      "Package": {
        "description": "An installed software package detected on the host.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "ecosystem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "OSV ecosystem for vulnerability matching, e.g. 'Debian:12', 'PyPI', 'npm'. When unset, detection falls back to the host's ecosystem derived from host.os.",
            "title": "Ecosystem"
          },
          "install_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Path"
          },
          "kind": {
            "const": "package",
            "default": "package",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Package manager, e.g. apt / yum / pip / npm",
            "title": "Source"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source package name when the package manager exposes it",
            "title": "Source Name"
          },
          "source_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source package version used to build the installed binary package",
            "title": "Source Version"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "name",
          "version"
        ],
        "title": "Package",
        "type": "object"
      },
      "Port": {
        "description": "A listening network port and the process bound to it.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "kind": {
            "const": "port",
            "default": "port",
            "title": "Kind",
            "type": "string"
          },
          "listen_addr": {
            "title": "Listen Addr",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "pid": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "port": {
            "maximum": 65535.0,
            "minimum": 0.0,
            "title": "Port",
            "type": "integer"
          },
          "process_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Name"
          },
          "proto": {
            "enum": [
              "tcp",
              "udp"
            ],
            "title": "Proto",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "proto",
          "port",
          "listen_addr"
        ],
        "title": "Port",
        "type": "object"
      },
...
      "SecurityProduct": {
        "description": "Endpoint security product and the protection state observed on a host.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "behavior_monitor": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behavior Monitor"
          },
          "cloud_protection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Protection"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "ioav_protection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ioav Protection"
          },
          "kind": {
            "const": "security_product",
            "default": "security_product",
            "title": "Kind",
            "type": "string"
          },
          "last_full_scan_at": {
            "anyOf": [
              {
                "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Full Scan At"
          },
          "last_quick_scan_at": {
            "anyOf": [
              {
                "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Quick Scan At"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vendor-specific running mode, for example Normal or Passive Mode",
            "title": "Mode"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "product_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Version"
          },
          "real_time_protection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Real Time Protection"
          },
          "signature_updated_at": {
            "anyOf": [
              {
                "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Updated At"
          },
          "signature_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Version"
          },
          "signatures_out_of_date": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signatures Out Of Date"
          },
          "status": {
            "enum": [
              "active",
              "passive",
              "disabled",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          },
          "tamper_protection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tamper Protection"
          },
          "vendor": {
            "title": "Vendor",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "name",
          "vendor",
          "status"
        ],
        "title": "SecurityProduct",
        "type": "object"
      },
      "Service": {
        "description": "A system service (daemon) and its current run state.",
        "properties": {
          "asset_id": {
            "description": "Stable identifier assigned by the scanner",
            "maxLength": 4096,
            "title": "Asset Id",
            "type": "string"
          },
          "exec_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exec Path"
          },
          "kind": {
            "const": "service",
            "default": "service",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent asset_id when this row came from a nested (container rootfs) scan",
            "title": "Parent Asset Id"
          },
          "status": {
            "description": "running / stopped / failed / ...",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "name",
          "status"
        ],
        "title": "Service",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Control plane and the only integration boundary for admin, analyzer, and agent.",
    "title": "kcatta Form",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ingest/asset-report": {
      "post": {
        "operationId": "ingest_asset_report_ingest_asset_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Asset Report"
      }
    },
...
  }
}